# apps/api/Program.cs
using Microsoft.AspNetCore.DataProtection;
using Microsoft.AspNetCore.HttpOverrides;
using System.Net;
using Microsoft.AspNetCore.RateLimiting;
using System.Security.Claims;
using System.Security.Cryptography;
using System.Text.Json;
using Microsoft.AspNetCore.Antiforgery;
using Microsoft.AspNetCore.Authentication;
using Microsoft.AspNetCore.Authentication.Cookies;
using Microsoft.AspNetCore.Identity;
using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
using Microsoft.EntityFrameworkCore;
using System.Threading.RateLimiting;

var builder = WebApplication.CreateBuilder(args);
var demo = builder.Configuration.GetValue("DemoMode", false);
var localHttp = builder.Configuration.GetValue<bool>("LocalHttp");
var connection = builder.Configuration.GetConnectionString("Postgres");
if (string.IsNullOrWhiteSpace(connection) && (!demo || !builder.Configuration.GetValue<bool>("AllowFileDemo"))) throw new InvalidOperationException("Configure PostgreSQL. The synthetic file adapter requires DemoMode=true and explicit AllowFileDemo=true.");
builder.Services.AddDataProtection().SetApplicationName("Wanees").PersistKeysToFileSystem(new DirectoryInfo(Path.Combine(builder.Environment.ContentRootPath, "data", "keys")));
builder.WebHost.ConfigureKestrel(o => o.Limits.MaxRequestBodySize = 65536);
builder.Services.Configure<ForwardedHeadersOptions>(o => { o.ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto; if (IPAddress.TryParse(builder.Configuration["TrustedProxy"], out var proxy)) o.KnownProxies.Add(proxy); });
builder.Services.AddProblemDetails();
builder.Services.AddOpenApi();
builder.Services.AddDbContext<WaneesDb>(o => o.UseNpgsql(connection ?? "Host=localhost;Database=wanees;Username=wanees;Password=configure-me"));
builder.Services.AddIdentityCore<IdentityUser>(o => { o.Password.RequiredLength = 12; o.User.RequireUniqueEmail = true; }).AddRoles<IdentityRole>().AddEntityFrameworkStores<WaneesDb>().AddSignInManager();
builder.Services.AddAuthentication("Wanees").AddCookie("Wanees", o => { o.Cookie.Name = "wanees-session"; o.Cookie.HttpOnly = true; o.Cookie.SameSite = SameSiteMode.Strict; o.Cookie.SecurePolicy = localHttp ? CookieSecurePolicy.SameAsRequest : CookieSecurePolicy.Always; o.ExpireTimeSpan = TimeSpan.FromHours(2); o.SlidingExpiration = false; o.Events.OnValidatePrincipal = Accounts.ValidateSession; o.Events.OnRedirectToLogin = c => { c.Response.StatusCode = 401; return Task.CompletedTask; }; o.Events.OnRedirectToAccessDenied = c => { c.Response.StatusCode = 403; return Task.CompletedTask; }; });
builder.Services.AddAuthorization(o => { o.AddPolicy("Guardian", p => p.RequireRole("Guardian")); o.AddPolicy("Staff", p => p.RequireRole("Editor", "Reviewer", "HospitalAdmin", "PlatformAdmin")); o.AddPolicy("Reviewer", p => p.RequireRole("Reviewer", "HospitalAdmin", "PlatformAdmin")); });
builder.Services.AddAntiforgery(o => { o.HeaderName = "X-CSRF-TOKEN"; o.Cookie.Name = "wanees-csrf"; o.Cookie.SameSite = SameSiteMode.Strict; o.Cookie.SecurePolicy = localHttp ? CookieSecurePolicy.SameAsRequest : CookieSecurePolicy.Always; });
builder.Services.AddRateLimiter(o => { o.RejectionStatusCode = 429; o.AddPolicy("auth", c => RateLimitPartition.GetFixedWindowLimiter(c.Connection.RemoteIpAddress?.ToString() ?? "anonymous", _ => new FixedWindowRateLimiterOptions { PermitLimit = 30, Window = TimeSpan.FromMinutes(1), QueueLimit = 0 })); });
builder.Services.AddRateLimiter(o => o.GlobalLimiter = PartitionedRateLimiter.Create<HttpContext, string>(c => RateLimitPartition.GetFixedWindowLimiter(c.Connection.RemoteIpAddress?.ToString() ?? "anonymous", _ => new FixedWindowRateLimiterOptions { PermitLimit = 300, Window = TimeSpan.FromMinutes(1), QueueLimit = 0 })));
builder.Services.AddHttpClient("models", c => c.Timeout = TimeSpan.FromSeconds(180));
builder.Services.AddRateLimiter(o => o.AddPolicy("chat", c => RateLimitPartition.GetFixedWindowLimiter(c.User.FindFirstValue(ClaimTypes.NameIdentifier) ?? c.Connection.RemoteIpAddress?.ToString() ?? "anonymous", _ => new FixedWindowRateLimiterOptions { PermitLimit = 12, Window = TimeSpan.FromMinutes(1), QueueLimit = 0 })));
if (demo && string.IsNullOrWhiteSpace(connection)) builder.Services.AddSingleton<IStore, DemoFileStore>(); else builder.Services.AddScoped<IStore, PgStore>();
var app = builder.Build();
app.UseExceptionHandler();
app.UseForwardedHeaders();
app.UseAuthentication();
app.Use(async (c, next) => { c.Response.Headers["X-Content-Type-Options"] = "nosniff"; c.Response.Headers["Referrer-Policy"] = "same-origin"; c.Response.Headers["Cache-Control"] = "no-store"; if (c.Request.Path.StartsWithSegments("/api") && c.Request.Method is not ("GET" or "HEAD" or "OPTIONS")) { try { await c.RequestServices.GetRequiredService<IAntiforgery>().ValidateRequestAsync(c); } catch (AntiforgeryValidationException) { await Results.Problem(statusCode: 400, title: "Invalid CSRF token").ExecuteAsync(c); return; } } await next(); });
app.UseAuthorization(); app.UseRateLimiter();
if (app.Environment.IsDevelopment()) app.MapOpenApi();
app.MapGet("/api/health", () => Results.Ok(new { status = "ok", mode = demo ? "synthetic-demo" : "live", storage = string.IsNullOrWhiteSpace(connection) ? "explicit-demo-file-adapter" : "postgresql" }));
var v = app.MapGroup("/api/v1");
v.MapGet("/session", (HttpContext c, IAntiforgery anti) => new { authenticated = c.User.Identity?.IsAuthenticated ?? false, role = c.User.FindFirstValue(ClaimTypes.Role), roles = c.User.FindAll(ClaimTypes.Role).Select(x => x.Value), email = c.User.FindFirstValue(ClaimTypes.Email), hospital = c.User.FindFirstValue("hospital"), demo, csrf = anti.GetAndStoreTokens(c).RequestToken });
v.MapPost("/demo/session", async (HttpContext c, DemoSession input) => { if (!demo) return Results.NotFound(); if (input.Role is not ("Guardian" or "Editor" or "Reviewer")) return Results.BadRequest(); var owner = Guid.NewGuid().ToString("N"); await Session(c, owner, input.Role, "al-bahar"); return Results.Ok(new { role = input.Role, synthetic = true }); }).RequireRateLimiting("auth");
Accounts.Map(v, demo);
CompanionChat.Map(v, demo);
v.MapGet("/hospitals", () => new[] { new { id = "al-bahar", name = "Al Bahar Children’s Centre", nameAr = "مركز البحار للأطفال", timeZone = "Asia/Muscat", fictional = true } });
v.MapGet("/pathways", () => new[] { new { id = "routine-xray", hospitalId = "al-bahar", ageBand = "5-10", version = "demo-1", status = "draft-demonstration", languages = new[] { "en", "ar" } } });
var guardian = v.MapGroup("/records").RequireAuthorization("Guardian");
guardian.MapGet("/{kind}", async (string kind, HttpContext c, IStore s) => !Kinds.Valid(kind) ? Results.NotFound() : Results.Ok(await s.List(kind, Owner(c))));
guardian.MapPost("/{kind}", async (string kind, RecordInput input, HttpContext c, IStore s) => { var error = Validate(kind, input.Data); if (error != null) return Results.Problem(statusCode: 400, title: error); if (!await ProfileOwned(input.Data, c, s)) return Results.NotFound(); var r = new OwnedRecord { Id = Guid.NewGuid(), Kind = kind, OwnerId = Owner(c), Data = input.Data.GetRawText() }; await s.Put(r); return Results.Created($"/api/v1/records/{kind}/{r.Id}", r); });
guardian.MapPut("/{kind}/{id:guid}", async (string kind, Guid id, RecordInput input, HttpContext c, IStore s) => { var r = await s.Get(id); if (r == null || r.OwnerId != Owner(c) || r.Kind != kind) return Results.NotFound(); var error = Validate(kind, input.Data); if (error != null) return Results.Problem(statusCode: 400, title: error); if (!await ProfileOwned(input.Data, c, s)) return Results.NotFound(); if (input.Version != r.Version) return Results.Problem(statusCode: 409, title: "This record has changed. Reload before saving."); r.Data = input.Data.GetRawText(); r.Version++; await s.Put(r); return Results.Ok(r); });
guardian.MapDelete("/{kind}/{id:guid}", async (string kind, Guid id, HttpContext c, IStore s) => { var r = await s.Get(id); if (r == null || r.OwnerId != Owner(c) || r.Kind != kind) return Results.NotFound(); await s.Delete(id); if (kind == "profiles") foreach (var child in await s.List(null, Owner(c))) if (child.Data.Contains(id.ToString())) await s.Delete(child.Id); return Results.NoContent(); });
v.MapGet("/account/export", async (HttpContext c, IStore s, WaneesDb db) => { var owner = Owner(c); return Results.Ok(demo ? await s.List(null, owner) : await db.Records.Where(r => r.OwnerId == owner).OrderBy(r => r.CreatedAt).ToListAsync()); }).RequireAuthorization();
v.MapPost("/shares", async (ShareInput input, HttpContext c, IStore s) => { var record = await s.Get(input.PassportId); if (record is null || record.OwnerId != Owner(c) || record.Kind != "passports") return Results.NotFound(); var token = Convert.ToHexString(RandomNumberGenerator.GetBytes(32)); var r = new OwnedRecord { Id = Guid.NewGuid(), OwnerId = Owner(c), Kind = "shares", Data = JsonSerializer.Serialize(new { passportId = input.PassportId, hash = Convert.ToHexString(SHA256.HashData(System.Text.Encoding.UTF8.GetBytes(token))), expires = DateTimeOffset.UtcNow.AddHours(1) }) }; await s.Put(r); return Results.Ok(new { id = r.Id, token, expiresInSeconds = 3600 }); }).RequireAuthorization("Guardian");
v.MapGet("/shared/{token}", async (string token, IStore s) => { if (token.Length != 64) return Results.NotFound(); var hash = Convert.ToHexString(SHA256.HashData(System.Text.Encoding.UTF8.GetBytes(token))); foreach (var r in await s.List("shares", null)) { using var d = JsonDocument.Parse(r.Data); var root = d.RootElement; if (root.GetProperty("hash").GetString() == hash && root.GetProperty("expires").GetDateTimeOffset() > DateTimeOffset.UtcNow) { var p = await s.Get(root.GetProperty("passportId").GetGuid()); return p != null ? Results.Ok(new { data = p.Data }) : Results.NotFound(); } } return Results.NotFound(); });
v.MapDelete("/shares/{id:guid}", async (Guid id, HttpContext c, IStore s) => { var r = await s.Get(id); if (r == null || r.OwnerId != Owner(c) || r.Kind != "shares") return Results.NotFound(); await s.Delete(id); return Results.NoContent(); }).RequireAuthorization("Guardian");
var staff = v.MapGroup("/staff").RequireAuthorization("Staff");
staff.MapGet("/content", async (HttpContext c, IStore s) => Results.Ok(await s.List("content", Hospital(c))));
staff.MapPost("/content", async (ContentInput input, HttpContext c, IStore s) => { if (Hospital(c) != input.HospitalId) return Results.Forbid(); if (input.SourceId is Guid sourceId) { var source = await s.Get(sourceId); if (source is null || source.Kind != "content" || source.OwnerId != Hospital(c)) return Results.NotFound(); } if (input.Language is not ("ar" or "en")) return Results.BadRequest(); if (string.IsNullOrWhiteSpace(input.Title) || input.Title.Length > 200 || input.Body.Length > 10000) return Results.BadRequest(); var r = new OwnedRecord { Id = Guid.NewGuid(), OwnerId = Hospital(c), Kind = "content", Data = JsonSerializer.Serialize(new { title = input.Title, body = input.Body, language = input.Language, status = "draft", author = Owner(c), reviewer = (string?)null, hospitalId = input.HospitalId, sourceId = input.SourceId }) }; await s.Put(r); await Audit(s, c, "content-created", r.Id); return Results.Ok(r); });
staff.MapPost("/content/{id:guid}/transition", async (Guid id, Transition input, HttpContext c, IStore s) => { var r = await s.Get(id); if (r == null || r.Kind != "content" || r.OwnerId != Hospital(c)) return Results.NotFound(); if (r.Version != input.Version) return Results.Conflict(); var d = JsonSerializer.Deserialize<Dictionary<string, JsonElement>>(r.Data)!; var old = d["status"].GetString(); var valid = (old, input.Status) switch { ("draft", "in-review") => true, ("in-review", "approved") => (c.User.IsInRole("Reviewer") || c.User.IsInRole("HospitalAdmin") || c.User.IsInRole("PlatformAdmin")), ("approved", "published") => (c.User.IsInRole("Reviewer") || c.User.IsInRole("HospitalAdmin") || c.User.IsInRole("PlatformAdmin")), ("in-review", "draft") => (c.User.IsInRole("Reviewer") || c.User.IsInRole("HospitalAdmin") || c.User.IsInRole("PlatformAdmin")), ("published", "retired") => (c.User.IsInRole("Reviewer") || c.User.IsInRole("HospitalAdmin") || c.User.IsInRole("PlatformAdmin")), _ => false }; if (!valid) return Results.Problem(statusCode: 403, title: "Review gate rejected the transition"); if (!demo && input.Status == "approved" && d["author"].GetString() == Owner(c)) return Results.Problem(statusCode: 403, title: "An independent reviewer is required"); d["status"] = JsonSerializer.SerializeToElement(input.Status); if (input.Status == "approved") { d["reviewer"] = JsonSerializer.SerializeToElement(Owner(c)); d["reviewedAt"] = JsonSerializer.SerializeToElement(DateTimeOffset.UtcNow); } r.Data = JsonSerializer.Serialize(d); r.Version++; await s.Put(r); await Audit(s, c, $"content-{input.Status}", id); return Results.Ok(r); });
staff.MapGet("/audit", async (HttpContext c, IStore s) => Results.Ok(await s.List("audit", Hospital(c))));
if (builder.Configuration.GetValue<bool>("MigrateOnStartup") && !string.IsNullOrWhiteSpace(connection)) { using var scope = app.Services.CreateScope(); await scope.ServiceProvider.GetRequiredService<WaneesDb>().Database.MigrateAsync(); }
if (!string.IsNullOrWhiteSpace(connection)) { using var scope = app.Services.CreateScope(); await Accounts.EnsureRoles(scope.ServiceProvider); }
if (args.Contains("--create-admin")) { using var scope = app.Services.CreateScope(); await Accounts.CreateFirstAdmin(scope.ServiceProvider); return; }
app.Run();
static string Owner(HttpContext c) => c.User.FindFirstValue(ClaimTypes.NameIdentifier)!;
static string Hospital(HttpContext c) => c.User.FindFirstValue("hospital") ?? "unassigned";
static async Task Session(HttpContext c, string id, string role, string? hospital) { var claims = new List<Claim> { new(ClaimTypes.NameIdentifier, id), new(ClaimTypes.Role, role) }; if (hospital != null) claims.Add(new("hospital", hospital)); await c.SignInAsync("Wanees", new ClaimsPrincipal(new ClaimsIdentity(claims, "Wanees"))); }
static async Task<bool> ProfileOwned(JsonElement data, HttpContext c, IStore s) { if (!data.TryGetProperty("profileId", out var p) || p.ValueKind == JsonValueKind.Null || (p.ValueKind == JsonValueKind.String && p.GetString() == "")) return true; if (p.ValueKind != JsonValueKind.String || !Guid.TryParse(p.GetString(), out var id)) return false; var r = await s.Get(id); return r?.Kind == "profiles" && r.OwnerId == Owner(c); }
static string? Validate(string kind, JsonElement data) { if (!Kinds.Valid(kind) || kind == "shares") return "Unsupported record kind"; if (data.ValueKind != JsonValueKind.Object || data.GetRawText().Length > 16000) return "Invalid record"; if (data.TryGetProperty("profileId", out var reference) && reference.ValueKind is not (JsonValueKind.Null or JsonValueKind.String)) return "Invalid profile reference"; if (kind == "profiles" && (!data.TryGetProperty("nickname", out var nick) || nick.ValueKind != JsonValueKind.String || string.IsNullOrWhiteSpace(nick.GetString()) || nick.GetString()!.Length > 40)) return "Nickname required, maximum 40 characters"; if (kind == "appointments") { if (!data.TryGetProperty("date", out var day) || day.ValueKind != JsonValueKind.String || !DateOnly.TryParseExact(day.GetString(), "yyyy-MM-dd", out _) || !data.TryGetProperty("time", out var time) || time.ValueKind != JsonValueKind.String || !TimeOnly.TryParseExact(time.GetString(), "HH:mm", out _)) return "A valid date and time are required"; } if (kind == "passports") { if (!data.TryGetProperty("reviewed", out var reviewed) || reviewed.ValueKind != JsonValueKind.True || !data.TryGetProperty("answers", out var answers) || answers.ValueKind != JsonValueKind.Array || answers.GetArrayLength() != 6) return "Reviewed preference answers required"; foreach (var answer in answers.EnumerateArray()) if (answer.ValueKind != JsonValueKind.Null && (answer.ValueKind != JsonValueKind.Number || !answer.TryGetInt32(out var n) || n < 0 || n > 2)) return "Invalid preference choice"; } return null; }
static async Task Audit(IStore s, HttpContext c, string action, Guid id) => await s.Put(new OwnedRecord { Id = Guid.NewGuid(), Kind = "audit", OwnerId = Hospital(c), Data = JsonSerializer.Serialize(new { action, recordId = id, actor = Owner(c), at = DateTimeOffset.UtcNow }) });
public record DemoSession(string Role);
public record Login(string Email, string Password);
public record Register(string Email, string Password, bool Consent);
public record RecordInput(JsonElement Data, int? Version);
public record ShareInput(Guid PassportId);
public record ContentInput(string HospitalId, string Title, string Body, string Language, Guid? SourceId = null);
public record Transition(string Status, int Version);
public static class Kinds { public static bool Valid(string k) => new[] { "profiles", "appointments", "passports", "checklists", "feelings", "notes", "consents", "shares" }.Contains(k); }
public class OwnedRecord { public Guid Id { get; set; } public string OwnerId { get; set; } = ""; public string Kind { get; set; } = ""; public string Data { get; set; } = "{}"; public int Version { get; set; } = 1; public DateTimeOffset CreatedAt { get; set; } = DateTimeOffset.UtcNow; }
public interface IStore { Task<List<OwnedRecord>> List(string? kind, string? owner); Task<OwnedRecord?> Get(Guid id); Task Put(OwnedRecord r); Task Delete(Guid id); }
public class WaneesDb(DbContextOptions<WaneesDb> options) : IdentityDbContext<IdentityUser>(options) { public DbSet<OwnedRecord> Records => Set<OwnedRecord>(); protected override void OnModelCreating(ModelBuilder b) { base.OnModelCreating(b); b.Entity<OwnedRecord>().HasIndex(x => new { x.OwnerId, x.Kind }); b.Entity<OwnedRecord>().Property(x => x.Data).HasColumnType("jsonb"); b.Entity<OwnedRecord>().Property(x => x.Version).IsConcurrencyToken(); } }
public class PgStore(WaneesDb db) : IStore { public Task<List<OwnedRecord>> List(string? kind, string? owner) => db.Records.Where(x => (kind == null || x.Kind == kind) && (owner == null || x.OwnerId == owner)).OrderByDescending(x => x.CreatedAt).Take(200).ToListAsync(); public Task<OwnedRecord?> Get(Guid id) => db.Records.SingleOrDefaultAsync(x => x.Id == id); public async Task Put(OwnedRecord r) { if (db.Entry(r).State == EntityState.Detached) { if (await db.Records.AnyAsync(x => x.Id == r.Id)) db.Records.Update(r); else db.Records.Add(r); } await db.SaveChangesAsync(); } public async Task Delete(Guid id) { await db.Records.Where(x => x.Id == id).ExecuteDeleteAsync(); } }
public class DemoFileStore : IStore { readonly string path; readonly SemaphoreSlim gate = new(1); public DemoFileStore(IConfiguration config) { path = Path.GetFullPath(config["DemoDataPath"] ?? "data/synthetic-demo.json"); Directory.CreateDirectory(Path.GetDirectoryName(path)!); } async Task<List<OwnedRecord>> Read() => File.Exists(path) ? JsonSerializer.Deserialize<List<OwnedRecord>>(await File.ReadAllTextAsync(path)) ?? [] : []; public async Task<List<OwnedRecord>> List(string? kind, string? owner) { await gate.WaitAsync(); try { return (await Read()).Where(x => (kind == null || x.Kind == kind) && (owner == null || x.OwnerId == owner)).OrderByDescending(x => x.CreatedAt).Take(200).ToList(); } finally { gate.Release(); } } public async Task<OwnedRecord?> Get(Guid id) { await gate.WaitAsync(); try { return (await Read()).Find(x => x.Id == id); } finally { gate.Release(); } } public async Task Put(OwnedRecord r) { await gate.WaitAsync(); try { var rows = await Read(); var old = rows.Find(x => x.Id == r.Id); if (old != null && r.Version != old.Version + 1) throw new DbUpdateConcurrencyException(); rows.RemoveAll(x => x.Id == r.Id); rows.Add(r); await File.WriteAllTextAsync(path + ".tmp", JsonSerializer.Serialize(rows)); File.Move(path + ".tmp", path, true); } finally { gate.Release(); } } public async Task Delete(Guid id) { await gate.WaitAsync(); try { var rows = await Read(); rows.RemoveAll(x => x.Id == id); await File.WriteAllTextAsync(path + ".tmp", JsonSerializer.Serialize(rows)); File.Move(path + ".tmp", path, true); } finally { gate.Release(); } } }
public partial class Program { }